refactor: moved everything into seperate files
