+++ projects/01_fyyur/starter_code/models/Show.py
@@ -1,0 +1,10 @@
+from models.shared import db
+
+
+class Show(db.Model):
+    __tablename__ = 'show'
+
+    id = db.Column(db.Integer, primary_key=True)
+    start_time = db.Column(db.String, nullable=False)
+    artist_id = db.Column(db.Integer, db.ForeignKey('artist.id'), nullable=False)
+    venue_id = db.Column(db.Integer, db.ForeignKey('venue.id'), nullable=False)
