+++ projects/01_fyyur/starter_code/models/shared.py
@@ -1,0 +1,3 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy();
